backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
python-multipart==0.0.6
python-dotenv==1.0.1
aiohttp==3.9.5
python-socketio==5.11.2